channel_simulator: gains for channels (for X2 handover with a COTS UE)
The main goal of this commit is to be able to perform X2 handover
using a COTS UE and two 'simulated' eNB, each running on a separate
computer, but both using the same USRP device to send and receive IQ
data over the air. By changin the gains at runtime using a small script,
it is possible to trigger X2 handover based on the A3 measurement report
as sent by the UE.

The documentation on how to perform X2 handover with only one
USRP device is in channel_simulator/README.x2_handover.